Created the Item.py class.
Description:
Item class has three attributes: name, value, and type, which are used to describe the item. The Item class also has four methods: __init__(), buy(), sell(), and transfer(), which are used to initialize the item, allow it to be bought and sold, and allow it to be transferred between subjects.